+++ Configurator.UI/ClientApp/src/components/FileUpload.js
@@ -60,31 +60,6 @@ export class FileUpload extends Component {
             });
     };
 
-    // Function to handle file download
-    handleDownload = (fileName) => {
-        fetch('upload/download/' + fileName, {
-            method: 'GET',
-        })
-            .then((response) => {
-                if (!response.ok) {
-                    throw new Error('Network response was not ok');
-                }
-                return response.blob();
-            })
-            .then((blob) => {
-                const url = window.URL.createObjectURL(blob);
-                const link = document.createElement('a');
-                link.href = url;
-                link.setAttribute('download', fileName);
-                document.body.appendChild(link);
-                link.click();
-            })
-            .catch((error) => {
-                console.error(error);
-                alert('File download failed.');
-            });
-    };
-
     // Fetch files from the server on component mount
     componentDidMount() {
         fetch('upload/files')
