+++ Configurator.UI/Controllers/DownloadController.cs
@@ -1,0 +1,12 @@
+using Microsoft.AspNetCore.Mvc;
+
+namespace Controller.UI.Controllers
+{
+    [ApiController]
+    [Route("[controller]")]
+
+    public class DownloadController : ControllerBase
+    {
+        
+    }
+}
